move login and change paths
