+++ src/components/Onboarding/SignUpForm.js
@@ -10,7 +10,7 @@ import {
   buttonTheme,
   Logo,
 } from '../Landing/Landing-styles';
-import { FormCard, FormContainer } from '../Forms/LoginForm';
+import { FormCard, FormContainer } from './LoginForm';
 
 import { register } from '../../state/actions/authenticationActions';
